modelo device

+++ EnergyMonitoringBack/backend/EnergyMonitoring/Application/DTOs/Divices/DeviceDTO.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
         public string? Location { get; set; }
-        public decimal? RatedPowerWatts { get; set; }
+        public double? RatedPowerWatts { get; set; }
         public decimal ExpectedMonthlyConsumptionKWh { get; set; }
         public bool IsActive { get; set; }
     }
